More refactoring package structure
