+++ app/src/main/java/widget/missions/MissionWidgetWorker.kt
@@ -1,4 +1,4 @@
-package widget
+package widget.missions
 
 import android.content.Context
 import androidx.work.CoroutineWorker
