+++ app/src/main/java/widget/missions/large/MissionWidgetLargeReceiver.kt
@@ -1,7 +1,7 @@
-package widget.large
+package widget.missions.large
 
 import androidx.glance.appwidget.GlanceAppWidget
-import widget.MissionWidgetReceiver
+import widget.missions.MissionWidgetReceiver
 
 class MissionWidgetLargeReceiver : MissionWidgetReceiver() {
     override val glanceAppWidget: GlanceAppWidget = MissionWidgetLarge()
