+++ app/src/main/java/widget/missions/minimal/MissionWidgetMinimalReceiver.kt
@@ -1,7 +1,7 @@
-package widget.minimal
+package widget.missions.minimal
 
 import androidx.glance.appwidget.GlanceAppWidget
-import widget.MissionWidgetReceiver
+import widget.missions.MissionWidgetReceiver
 
 class MissionWidgetMinimalReceiver : MissionWidgetReceiver() {
     override val glanceAppWidget: GlanceAppWidget = MissionWidgetMinimal()
